+++ src/main/java/fitnus/command/HelpCommand.java
@@ -1,9 +1,8 @@
 package fitnus.command;
 
-import fitnus.EntryDatabase;
-import fitnus.FitNusException;
-import fitnus.FoodDatabase;
-import fitnus.User;
+import fitnus.database.EntryDatabase;
+import fitnus.database.FoodDatabase;
+import fitnus.utility.User;
 
 public class HelpCommand extends Command {
     public HelpCommand() {
