+++ src/main/java/fitnus/command/ListFoodIntakeCommand.java
@@ -1,9 +1,8 @@
 package fitnus.command;
 
-import fitnus.EntryDatabase;
-import fitnus.FoodDatabase;
-import fitnus.Ui;
-import fitnus.User;
+import fitnus.database.EntryDatabase;
+import fitnus.database.FoodDatabase;
+import fitnus.utility.User;
 
 public class ListFoodIntakeCommand extends Command {
     private final String timeSpan;
