+++ src/main/java/fitnus/exception/FitNusException.java
@@ -1,4 +1,4 @@
-package fitnus;
+package fitnus.exception;
 
 public class FitNusException extends Exception {
     public FitNusException(String exception) {
