+++ src/main/java/fitnus/utility/Storage.java
@@ -1,4 +1,8 @@
-package fitnus;
+package fitnus.utility;
+
+import fitnus.database.EntryDatabase;
+import fitnus.database.FoodDatabase;
+import fitnus.exception.FitNusException;
 
 import java.io.BufferedReader;
 import java.io.FileInputStream;
