+++ src/main/java/fitnus/utility/User.java
@@ -1,12 +1,11 @@
-package fitnus;
+package fitnus.utility;
 
-import fitnus.parser.Parser;
+import fitnus.database.EntryDatabase;
+import fitnus.exception.FitNusException;
+import fitnus.utility.Ui;
 
 import java.io.BufferedReader;
 import java.io.IOException;
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Date;
 
 public class User {
     private static final int MALE = 0;
