+++ src/test/java/fitnus/DatabaseTest.java
@@ -1,5 +1,7 @@
 package fitnus;
 
+import fitnus.database.FoodDatabase;
+import fitnus.exception.FitNusException;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
