+++ src/test/java/fitnus/FoodDatabaseTest.java
@@ -1,5 +1,8 @@
 package fitnus;
 
+import fitnus.database.FoodDatabase;
+import fitnus.exception.FitNusException;
+import fitnus.tracker.Food;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
