+++ src/test/java/fitnus/command/SetGenderCommandTest.java
@@ -1,9 +1,8 @@
 package fitnus.command;
 
-import fitnus.EntryDatabase;
-import fitnus.Food;
-import fitnus.FoodDatabase;
-import fitnus.User;
+import fitnus.database.EntryDatabase;
+import fitnus.database.FoodDatabase;
+import fitnus.utility.User;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
